20230517_File: board 패키지 생성 후 파일들 이동
