+++ src/main/frontend/src/page/board/QuestionDetail.js
